Tag flicker tests and only run them when environment variable SCALATEST_RUN_FLICKER_TESTS is set to true.
