+++ jvm/scalatest-test/src/test/scala/org/scalatest/prop/GeneratorSpec.scala
@@ -21,6 +21,7 @@ import scala.collection.immutable.SortedSet
 import scala.collection.immutable.SortedMap
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
+import org.scalatest.tagobjects.Flicker
 
 class GeneratorSpec extends AnyFunSpec with Matchers {
   describe("A Generator") {
@@ -2539,7 +2540,7 @@ class GeneratorSpec extends AnyFunSpec with Matchers {
         orCanon.toList should contain theSameElementsAs((gCanon.map(Good(_)) ++ bCanon.map(Bad(_))).toList)
       }
 
-      it("should produce an appropriate mix of Good and Bad") {
+      it("should produce an appropriate mix of Good and Bad", Flicker) {
         import Generator._
         import org.scalactic._
         val gen = orGenerator[Int, String]
@@ -2550,7 +2551,7 @@ class GeneratorSpec extends AnyFunSpec with Matchers {
         }
 
         // It's arbitrary, but we know that it produces Bad about a quarter of the time:
-        classification.percentages("Bad").value should be (25 +- 2)
+        classification.percentages("Bad").value should be (25 +- 3)
       }
 
       it("should use the base types to shrink") {
